TapHandler manual test: demonstrate tapped signal
Missiles are too expensive to launch in continuous sprays anyway.
Also, it's too hard to keep holding down the "balloons" button if
we use policy DragThreshold.  This is better to have on the
fighters button, since the usual use case is that ballons are
launched first, and cheap enough to launch continuously.

Change-Id: I3b52556b81afad9fb7ec1a4b1dec4dde3bab104c
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>