Windows: Fix distance field rendering with fonts with broken cmap
If the CMAP table of a font is not present on Windows, then the
font will not be counted as a ttf file, likewise if the CMAP table
has a bug in it. An effect of this is that the glyphCount() will
be returned as 0 since we will bail out before fetching the maxp
table, and the distance field renderer assumes the glyph count is
valid.

While the font in question is obviously broken and may have other
issues, it is possible to still support using it for rendering text
by bypassing the check when we cannot get a reliable glyph count from
the font.

[ChangeLog][Windows][Text] Worked around an issue with fonts that
have corrupt cmap tables.

Task-number: QTBUG-45321
Change-Id: I9cfd7809661ae6b902f6b9d47ffb18bfbc06bcd4
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>