QQuickPopup: use QShortcutMap to grab Back & Escape keys
This makes CloseOnEscape play well together with QML Shortcut, as
illustrated by the Gallery example.

For example, navigate to any control page, open the options menu,
and hit escape twice. First, the options menu closes, and then the
stackview navigates back.

Another nice example is to open the style combobox in the settings
dialog, while being navigated to one of the control pages, and hit
escape three times. First, the combobox closes, then the settings
dialog closes, and finally, the stackview navigates back.

NOTE: tst_combobox had to be updated, because it was assuming that
      the popup is closed when escape is released. Now that the
      system is based on shortcut overrides, it gets closed on
      press instead.

Task-number: QTBUG-56562
Change-Id: I8b8901bcba7deebd82b181af42f335d95a7cb469
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>