Gallery: add Escape shortcut
This does not work very well, yet. The purpose is to demonstrate the
lack of co-operation between of QML Shortcuts vs. QQC2 Popups. For
example, the back navigation shortcut activates in the background
when the modal settings dialog is open. The consequent commits are
going to fix these issues step by step.

Task-number: QTBUG-56562
Change-Id: Iceafc175f4da8054b65d664a3e0faf0087d0dec5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>