Universal: fix the size of a footer-less dialog
The Gallery example's About dialog is broken in the Universal style.
The dialog doesn't have any bottom padding.

Dialog cannot have bottom padding hard-coded to 0. It was assuming
that there would be a DialogButtonBox that provides top padding. Set a
suitable padding in the dialog instead, so that the dialog content gets
padded without a footer. Due to this, we must tweak the top and bottom
paddings of DialogButtonBox so that the total spacing between the
dialog content and the buttons becomes 24 as it is in the specs.

Change-Id: I77e2c3f66c5166391f9be595121f23bc95debde8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>