Fix popups to close on click outside
When a popup closes itself on press outside, it accepts the press
event so that it doesn't propagate to other popups or the content
below. We must make sure that such closing popup does not become
the mouse grabber, because it doesn't make sense to route the
subsequent mouse events to the popup that was just closed.

Change-Id: I80c6e26a1d94aa1526a61862f00af2fd0778aa82
Task-number: QTBUG-56697
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>