Replace QStringLiteral with QLatin1String in the proxy themes
This allows us, in theory, to test multiple styles in the same process
by calling qmlClearTypeRegistrations() between each style test round.
When QML types are cleared, QML plugins are unloaded, and the current
style plugin's proxy platform theme is destroyed. It leads to a nasty
crash if there are fonts stored in the Qt font database that have their
family names stored in the read-only data of an unloaded plugin.

Change-Id: I32a70077ac4502e16a753fb68f7efeef35d239b5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>