Change Import::type from quint8 to LEUInt32
Import::type used to be a quint32.  The switch from quint32 to quint8
changed the alignment of other structures in the compiled blob.  On QNX,
the alignment isn't appropriate for some of the instructions that the
compiler uses to access/modify the structures and a BUS error occurs.
Changing Import::type to LEUInt32 puts the alignment back the way it
was before.  Simpler than trying to figure out how to get everything
aligned correctly with a quint8 at the front of the structure.

Task-number: QTBUG-56341
Change-Id: Ia6f4daff5f615f9c5dbcc34ce8c3c9a89bab2b69
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>