ComboBox: prepare the internal layouts for editable combos
The text field of an editable ComboBox must cover the whole control
excluding the indicator button. This ensures that user interaction
works as expected over the whole editor area (there must be no "dead"
paddings), and that the whole editor area gets the appropriate caret
or ibeam mouse cursor.

Previously, the indicator was laid out inside/over the content item.
To help with getting the layout right, this change injects empty
padding into the indicator images of the Universal and Material styles
(the Default style already had padding).

Task-number: QTBUG-53876
Change-Id: I62e10ffb7f5ed4559523513f04ff7e2ea0e57ef5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>