Default: prepare ComboBox for deferred popup execution
The plan is to defer the QML execution of the whole ComboBox popup
until the popup is made visible or accessed by the user. This can
slice off roughly half of the creation time.

The QML engine refuses to defer the execution if there are any child
objects that have an identifier, because those child objects could be
referenced from outside the deferred/missing parent object. For this
reason, we must kill the unnecessary ID. This change alone does not
defer anything, but is a prerequisite for enabling deferred execution
later.

Flickable overrides Item's default property to flickableData, which
ensures that children get re-parented to Flickable::contentItem, but
ListView overrides the default property back to Item::data, so the
parent assignment is not necessary at all. The rest can be done by
referencing parent instead of the ID.

The other styles do not have ID's within the popup.

Change-Id: Iac58bae7e99252fb03bed2228153483300901720
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>