Add QQuickComboBox::down
This allows us to replace all "control.pressed || popup.visible"
expressions with "control.down", and takes us one step closer to
deferred popop execution as mentioned in the previous commit. In
order to be able to defer the popup execution, we must get rid of
such explicit references to the popup.

Change-Id: Ifa7ecf8201912d3ec1bca232e2cf600e3886715e
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>