QQuickComboBox: ensure a parent for delegates
This allows us to replace the "width: control.popup.width" expressions
with "width: parent.width" to avoid explicit references to the popup.
Next, we are ready to defer the execution of the popup until it becomes
visible or is accessed by the user.

Without the parent item fix, ComboBox throws an error:

    ComboBox.qml:58: TypeError: Cannot read property of null

...and tst_snippets fails => tested implicitly. :)

Change-Id: Ib9cfcf7558addbe3403d835d3ae11b6c11c2f0cb
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>