tst_qqmllocale: use unsetenv to clear environment variable
tst_qqmllocale::timeZoneUpdated() was recording the prior time zone,
frobbing it for the duration of its test, then restoring it.  However,
if TZ was previously unset, it got set to empty instead of being
unset.  This did not quite have the same effect; UTC ends up being
used instead of the system default, with potential for conflict
between system functions (that have attended to tzset) and third-party
libraries (e.g. ICU) with internal state saved from before the change.

Change-Id: I9f2147cd8858316e9ba5fd84d95dfc2c4538b0ab
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>