Stabilize tst_drawer
The test was using QTRY_COMPARE to test that the logical position of
the drawer was 1.0, meaning that the drawer was assumed to be fully
open. Then the test continued with touch move events and tested the
logical position accordingly.

The problem was that waiting (QTRY_COMPARE) for the logical position
of 1.0 does not ensure that the open transition has been finished.
Fuzzy compare might return true just before the transition finishes.
We sent touch move events and tested the logical position accordingly,
while the open transition was still running. This way, the transition
ended up setting an unexpected position for the drawer in the middle
of the touch move event tests.

The solution is to wait for the opened() signal to ensure that the
transition has been finished, before continuing to test touch move
events.

Change-Id: Ia48f4cedc97c09bb1ee064f3b535ad4fc7ae5c71
Task-number: QTBUG-56061
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>