QQuickAbstractButton: fix auto-repeat
Don't cancel auto-repeat on the tiniest mouse/touch move, but keep
repeating until moved outside the button.

The test has been changed so that the exact amount of repeats does
not matter, as long as it repeats. This is because waits are not
reliable in a busy CI environment. Sometimes waits can take longer,
timer events get queued, and we get an unexpected burst of repeats.

Change-Id: Ic473e04c4d15a0826c8adf460c69507e64590d99
Task-number: QTBUG-57085
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>