Restore AbstractButton::checkable
In 7a8f055, we hid the checkable property from QQuickAbstractButton and
exposed it only in QQuickButton and QQuickMenuItem. The reasoning was
that QQuickCheckBox, QQuickRadioButton, QQuickSwitch and their delegate
counterparts are inherently checkable, so having the checkable property
available in QML didn't seem to make sense.

While this still holds true, there are other factors to take into
consideration. AbstractButton is meant to be a generic base class for
all types of buttons, but the lack of a checkable property makes it
unusable as a base class for custom QML-based checkable buttons.

If we want to a hide the checkable property from CheckBox, RadioButton,
and Switch to avoid having it popup in the auto-completion list, we can
probably do it in a less disruptive way via tooling.

[ChangeLog][Controls][AbstractButton] The checkable property has been
made accessible from QML. Previously it was only exposed for Button and
MenuItem, but it is now available for any AbstractButton to make it
possible to create custom QML-based checkable buttons.

Task-number: QTBUG-51554
Change-Id: I19e29fc87cd15811c43c9b9ebb29701d66cde72f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>