Material: de-couple foreground and primaryTextColor
In order to get the ComboBox colors right, it must be possible to query
the original primary text color value. It was very convenient to make
Material.primaryTextColor return custom Material.foreground colors,
because that way we didn't have to touch our QML files, but this
approach is no longer usable.

Task-number: QTBUG-57167
Change-Id: Iee003e96112b919dc7c84a906e4f5691b0f2a6ab
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>