QQuickStyle: kill the temporary QQmlEngine instance
The temporary QQmlEngine instance is not only slow and ugly, but also
causes random crashes in the CI. This change replaces the temporary
QQmlEngine instance with a local defaultImportPathList() helper method,
which is based on the code in QQmlImportDatabase constructor. Later on,
we can switch to QQmlImportDatabase::defaultImportPathList() when it
has been made available in dev.

Change-Id: I9f8363ba79ce39fb0482e6e44be0b1d1040c76e4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>