deliverUpdatedTouchPoints: if a handler is the grabber, it goes first
deliverMatchingPointsToItem() does not discriminate between handlers:
it delivers to them in the order they are in the item's vector.
In the map.qml manual test, this meant that DragHandler could steal
the grab even when the pinch is active.  On one hand, always delivering
to all handlers could help them to cooperate better.  On the other
hand, since we decided to let handlers grab directly (instead of only
allowing items to grab), that should be respected.  And DragHandler
so far does not refuse to allow dragging when two fingers are pressed
inside the parent item: it will grab one point and ignore the other.

Change-Id: I245eceafaea96096030ba6ff7263d95069e286ba
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>