Sort out and clarify QML type registration
We cannot call qmlRegisterRevision() for our own types, because that
will register an unnamed type. In practice, the empty type name occurs
as a missing type name in qmlInfo() warnings. I've clarified the type
registration comments a bit, about what needs to be registered and
where:

- templates: new types _and_ new revisions for existing types
- controls: _only_ new types

For controls, the revisions is determined by the import version in
the .qml file, so those must be kept up to date in new releases.

Change-Id: I7c76a9d80185fa7e922c419b4ee7361be45c61f5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>