SwipeDelegate: fix swipes that begin over a child item of a delegate
If the mouse was pressed over a child item of the delegate, the event's
position will be for that item, rather than the entire control. To
fix this, we need to set the correct pressPoint. To avoid duplicating
QQuickAbstractButton::mousePressEvent()'s code, we simply set the
correct pressPoint after calling it in
QQuickSwipeDelegatePrivate::handleMousePressEvent().

Task-number: QTBUG-57271
Change-Id: I7204d6176c755512eeeb0b1a57a920fbab40392c
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>