SwipeDelegate: transition swipe.position
Previously the value of swipe.position jumped on mouse/touch release,
or when open() or close() was called. Transitioning the swipe.position
value smoothly to the target value makes it possible to use it as
opacity/scale/etc. for the side items.

Notice that this is backwards compatible with the old behavior. If no
transition is set, the value jumps the same way it did before. In that
case, Behavior on x makes the movement smooth.

Task-number: QTBUG-57242
Change-Id: Id9c06b5b08fa73f2f575787e16dc6e20e4ccb545
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>