Allow specifying a relative style path in qtquickcontrols2.conf
The style path is resolved relative to the location of config file.
Even though qtquickcontrols2.conf is normally located in the root of
resources, Quick Designer uses QT_QUICK_CONTROLS_CONF to set a custom
location for the preview, where the standard config location is not
feasible.

This patch extends the config and style lookup further by allowing
to specify a relative path to a custom style for the Quick Designer
preview.

Change-Id: Iad9c459421eefbf7a5bcbedfea59f5f16a1d3a39
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>