Allow specifying a relative style path in qtquickcontrols2.conf
The style path is resolved relative to the location of config file.
Even though qtquickcontrols2.conf is normally located in the root of
resources, Quick Designer uses QT_QUICK_CONTROLS_CONF to set a custom
location for the preview, where the standard config location is not
feasible.

This patch extends the config and style lookup further by allowing
to specify a relative path to a custom style for the Quick Designer
preview.

This is a cherry-pick of 5efd4b6a61498f0668d7367d620275367894140e,
which fixes the style lookup so that QTBUG-57618 gets resolved.

Change-Id: I7b656c42cb6aca0914ad20c94c7b527bea74cb16
Task-number: QTBUG-57618
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>