SwipeDelegate: fix animations when releasing from a drag
9812a9c changed the order in which the pressed and position
state is set. Before, when releasing after swiping, it would be:

set pressed to false => set position

After 9812a9c, it became:

set position => set pressed to false

The original order is necessary to ensure that animations can rely
on being enabled *before* position changes, as their enabled
expression typically looks something like this:

    enabled: !control.down

This patch duplicates the contents of
QQuickAbstractButton::mouseUngrabEvent() for now, with a TODO
comment to ensure that it's moved into a private helper that we can
call later on.

Task-number: QTBUG-57350
Change-Id: I31af7a665fb2d0e37548df31560ed7bbb0c3cadb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>