V4: Prevent repeatedly loading a QV4::Function pointer
Even with LTO on, GCC6 will not optimize out the fact that a) the d()
pointer won't change here, and that d()->function is constant and not
need to be checked for nullptr all the time.

Also prevents allocating 0 locals, which would still run most of the
jsAlloca code for no good reason.

Change-Id: Iaf5257743cb57d55bef5d334d1e4355590dae7c6
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>