Add QQuickSpinBox::valueModified()
Another name candidate was valueEdited(), but it was left available
in case we want to have the respective signal for text editing in the
future.

[ChangeLog][Controls][SpinBox] Added a valueModified() signal that
is emitted whenever the value of a spin box has been interactively
modified by the user by using either touch, mouse, wheel, or keys.

Task-number: QTBUG-57203
Change-Id: I705c7e63d23235f51d401abf27f3458f8a5b0589
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>