V4 Interpreter: reset the exception handler for every catch block
If an exception handler consists of multiple basic-blocks, and is then
followed by a finally block (for example: a catch block with a throw
followed by a return, and then a subsequent finally block), then the
finally block can be reached before the exception handler is reset by
skipping the second block of the catch handler with a throw. As the
finally block will then rethrow, and the old exception handler is still
there, it will end in an endless loop.

The fix is to reset the exception handler for every catch block. The
problem occurred in ch12/12.14/S12.14_A13_T3

Change-Id: I968b62c6c2af30af32e2717a9ce85f852523dbe8
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>