QQuickWindowPrivate::dragOverThreshold(): ignore zero velocity threshold
Previously, this would trigger a drag even if the touch point was quite
steady, since a rather steady finger could report a very small velocity
(usually between 0 and 1).

Change so that it will ignore velocity if it's not positive.

At the same time convert it to a template function since we want to also
use this for QQuickEventPoint later on.

Change-Id: Ibb2210813707399ae84e3422718c995897891060
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>