QQuickSpriteEngine: Optimistic speedup for updateSprites
Switch from QList to QVector, and avoid using pop_front -- remove all in one
operation instead.

The idea here was to provide some gain to sprite-sequence in qmlbench, but any
possible gain is pretty miniscule due to the lack of batching that example
suffers. Still, we end up with less CPU time taken, which is a win of sorts.

Change-Id: Ib7c101afa71ee617c3c4506b14701d6a4143ed3a
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>