Material: prevent ripples from getting stuck in disabled state
Ripples don't trigger when the ripple item itself is disabled. This is
used in SwipeDelegate to prevent ripples when a swipe is open. Now that
swipe.position is transitioned, a release event comes before the swipe
position reaches 0, so we must ensure that the wave exits even if it's
not yet enabled.

Change-Id: Ib4d58ff974262331898db2133e06624eb890f9ed
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>